refactor: rename GraphProvider.KUZU to LADYBUG, kuzu/ to ladybug/
Rename the driver directory, all Kuzu*Operations classes to
Ladybug*Operations, and update the GraphProvider enum. KuzuDB was
abandoned upstream — LadybugDB is the active fork.
