+++ graphiti_core/driver/ladybug/operations/has_episode_edge_ops.py
@@ -17,22 +17,21 @@ limitations under the License.
 import logging
 from typing import Any
 
-from graphiti_core.driver.driver import GraphProvider
-from graphiti_core.driver.operations.community_edge_ops import CommunityEdgeOperations
+from graphiti_core.driver.operations.has_episode_edge_ops import HasEpisodeEdgeOperations
 from graphiti_core.driver.query_executor import QueryExecutor, Transaction
-from graphiti_core.edges import CommunityEdge
+from graphiti_core.edges import HasEpisodeEdge
 from graphiti_core.errors import EdgeNotFoundError
 from graphiti_core.helpers import parse_db_date
 from graphiti_core.models.edges.edge_db_queries import (
-    COMMUNITY_EDGE_RETURN,
-    get_community_edge_save_query,
+    HAS_EPISODE_EDGE_RETURN,
+    HAS_EPISODE_EDGE_SAVE,
 )
 
 logger = logging.getLogger(__name__)
 
 
-def _community_edge_from_record(record: Any) -> CommunityEdge:
-    return CommunityEdge(
+def _has_episode_edge_from_record(record: Any) -> HasEpisodeEdge:
+    return HasEpisodeEdge(
         uuid=record['uuid'],
         group_id=record['group_id'],
         source_node_uuid=record['source_node_uuid'],
@@ -41,36 +40,45 @@ def _community_edge_from_record(record: Any) -> CommunityEdge:
     )
 
 
-class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
+class LadybugHasEpisodeEdgeOperations(HasEpisodeEdgeOperations):
     async def save(
         self,
         executor: QueryExecutor,
-        edge: CommunityEdge,
+        edge: HasEpisodeEdge,
         tx: Transaction | None = None,
     ) -> None:
-        query = get_community_edge_save_query(GraphProvider.KUZU)
         params: dict[str, Any] = {
-            'community_uuid': edge.source_node_uuid,
-            'entity_uuid': edge.target_node_uuid,
+            'saga_uuid': edge.source_node_uuid,
+            'episode_uuid': edge.target_node_uuid,
             'uuid': edge.uuid,
             'group_id': edge.group_id,
             'created_at': edge.created_at,
         }
         if tx is not None:
-            await tx.run(query, **params)
+            await tx.run(HAS_EPISODE_EDGE_SAVE, **params)
         else:
-            await executor.execute_query(query, **params)
+            await executor.execute_query(HAS_EPISODE_EDGE_SAVE, **params)
 
         logger.debug(f'Saved Edge to Graph: {edge.uuid}')
+
+    async def save_bulk(
+        self,
+        executor: QueryExecutor,
+        edges: list[HasEpisodeEdge],
+        tx: Transaction | None = None,
+        batch_size: int = 100,
+    ) -> None:
+        for edge in edges:
+            await self.save(executor, edge, tx=tx)
 
     async def delete(
         self,
         executor: QueryExecutor,
-        edge: CommunityEdge,
+        edge: HasEpisodeEdge,
         tx: Transaction | None = None,
     ) -> None:
         query = """
-            MATCH (n:Community)-[e:HAS_MEMBER {uuid: $uuid}]->(m)
+            MATCH (n:Saga)-[e:HAS_EPISODE {uuid: $uuid}]->(m:Episodic)
             DELETE e
         """
         if tx is not None:
@@ -87,7 +95,7 @@ class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
         tx: Transaction | None = None,
     ) -> None:
         query = """
-            MATCH (n:Community)-[e:HAS_MEMBER]->(m)
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
             WHERE e.uuid IN $uuids
             DELETE e
         """
@@ -100,16 +108,16 @@ class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
         self,
         executor: QueryExecutor,
         uuid: str,
-    ) -> CommunityEdge:
+    ) -> HasEpisodeEdge:
         query = (
             """
-            MATCH (n:Community)-[e:HAS_MEMBER {uuid: $uuid}]->(m)
+            MATCH (n:Saga)-[e:HAS_EPISODE {uuid: $uuid}]->(m:Episodic)
             RETURN
             """
-            + COMMUNITY_EDGE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
         )
         records, _, _ = await executor.execute_query(query, uuid=uuid)
-        edges = [_community_edge_from_record(r) for r in records]
+        edges = [_has_episode_edge_from_record(r) for r in records]
         if len(edges) == 0:
             raise EdgeNotFoundError(uuid)
         return edges[0]
@@ -118,17 +126,17 @@ class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
         self,
         executor: QueryExecutor,
         uuids: list[str],
-    ) -> list[CommunityEdge]:
+    ) -> list[HasEpisodeEdge]:
         query = (
             """
-            MATCH (n:Community)-[e:HAS_MEMBER]->(m)
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
             WHERE e.uuid IN $uuids
             RETURN
             """
-            + COMMUNITY_EDGE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
         )
         records, _, _ = await executor.execute_query(query, uuids=uuids)
-        return [_community_edge_from_record(r) for r in records]
+        return [_has_episode_edge_from_record(r) for r in records]
 
     async def get_by_group_ids(
         self,
@@ -136,19 +144,19 @@ class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
         group_ids: list[str],
         limit: int | None = None,
         uuid_cursor: str | None = None,
-    ) -> list[CommunityEdge]:
+    ) -> list[HasEpisodeEdge]:
         cursor_clause = 'AND e.uuid < $uuid' if uuid_cursor else ''
         limit_clause = 'LIMIT $limit' if limit is not None else ''
         query = (
             """
-            MATCH (n:Community)-[e:HAS_MEMBER]->(m)
+            MATCH (n:Saga)-[e:HAS_EPISODE]->(m:Episodic)
             WHERE e.group_id IN $group_ids
             """
             + cursor_clause
             + """
             RETURN
             """
-            + COMMUNITY_EDGE_RETURN
+            + HAS_EPISODE_EDGE_RETURN
             + """
             ORDER BY e.uuid DESC
             """
@@ -160,4 +168,4 @@ class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
             uuid=uuid_cursor,
             limit=limit,
         )
-        return [_community_edge_from_record(r) for r in records]
+        return [_has_episode_edge_from_record(r) for r in records]
